Remove unecessary complexity
The original Bosto-2G driver is a fork of the existing Hanwang tablet
which already exists in the Linux Kernel, with the original license
removed, and all instances of `hanwang` replaced with `bosto_2g`.

The only necessary changes to the Hanwang Driver are to include the USB
deviceId for the Bosto2g.

This commit removes all other useless code/scripts/garabge which was
making it look like it actually did anything.