Attempt to compare Bosto2G to Hanwang driver
This commit establishes a direct comparison of the Bosto2G tablet driver
with the Hanwang tablet driver found in the Linux kernel.  The general
conclusion is that the vast majority of the Bosto2G driver is a simple
sed string replace of hanwang with bosto2g with the exception of a chunk
of code in the input stream parser being refactored in the bosto2g
driver.  It is not yet clear if this refactoring was actually adding
useful code or if the other was attempting to simplify what they felt
was a confusing state case.