Re-add detach_usbhid.c program
The Bosto line of tablet device ids fall outside of the range of the
first/last Hanwang device identifiers.  As such, the Linux USB HID layer
will attach to the the Bosto devices instead of the hanwang driver.

In order to test the driver we are resurecting the old detach_usbhid
from the old bosto2g code tree.